// src/modules/adoption-request/enums/rejection-reason.enum.ts
export enum RejectionReason {
  NOT_ENOUGH_EXPERIENCE = 'NOT_ENOUGH_EXPERIENCE',
  INSUFFICIENT_LIVING_SPACE = 'INSUFFICIENT_LIVING_SPACE',
  NO_YARD_FOR_PET = 'NO_YARD_FOR_PET',
  FINANCIAL_UNSTABLE = 'FINANCIAL_UNSTABLE',
  COMMITMENT_QUESTIONS = 'COMMITMENT_QUESTIONS',
  OTHER_PETS_INCOMPATIBLE = 'OTHER_PETS_INCOMPATIBLE',
  PET_ALREADY_ADOPTED = 'PET_ALREADY_ADOPTED',
  INCOMPLETE_APPLICATION = 'INCOMPLETE_APPLICATION',
  OTHER_REQUEST_APPROVED = 'OTHER_REQUEST_APPROVED',
}